fix(main): portable statvfs multiplication on macOS x86_64
nix::sys::statvfs::Statvfs::blocks_available() returns u64 on
Linux and macOS ARM, but u32 on macOS x86_64 (legacy fsblkcnt_t
ABI). The u32 * u64 multiply in available_disk_space_bytes
fails to compile for the x86_64-apple-darwin target with
E0277/E0308.

Fix by widening both operands to u64 via 'as u64' before the
multiply. This is a no-op on platforms where the type is
already u64, and a widening cast on macOS x86_64.

+++ src/main.rs
@@ -2736,7 +2736,10 @@ fn format_human_duration(secs: u64) -> String {
 fn available_disk_space_bytes(path: &std::path::Path) -> Option<u64> {
     use nix::sys::statvfs::statvfs;
     let stat = statvfs(path).ok()?;
-    Some(stat.blocks_available() * stat.block_size())
+    // blocks_available() / block_size() have different integer widths across
+    // platforms (both u64 on Linux and macOS ARM; blocks_available is u32 on
+    // macOS x86_64). Widen both to u64 so the multiply compiles everywhere.
+    Some(stat.blocks_available() as u64 * stat.block_size() as u64)
 }
 
 #[cfg(not(unix))]
